requirements.txt
python-telegram-bot==21.9
openai==1.57.0
requests==2.32.3
pytz==2024.2
python-dotenv==1.0.1
pillow==10.4.0
aiohttp==3.11.10